requirements.txt
fastapi
uvicorn
openai-whisper
neo4j
chromadb
requests
pydantic
python-dotenv
python-multipart
openai
